requirements.txt
flask>=2.0.0
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.22.0
pathlib>=1.0.1
mysql-connector-python
mysqlclient
flask-sqlalchemy
plotly>=5.0.0
pandas>=1.3.0
Werkzeug>=2.0.0